added flush-thread for index cache and added language-name mapping in Language_p
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@203 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/kelondro/kelondroMap.java
@@ -173,9 +173,9 @@ public class kelondroMap {
                 valuel = Long.parseLong(value);
                 accumulator = (Long) accMap.get(accfields[i]);
                 if (add)
-                    accMap.put(accfields[i], new Long(accumulator.longValue() + valuel));
+                    accMap.put(accfields[i], new Long(accumulator.longValue() + ((long) valuel)));
                 else
-                    accMap.put(accfields[i], new Long(accumulator.longValue() - valuel));
+                    accMap.put(accfields[i], new Long(accumulator.longValue() - ((long) valuel)));
             } catch (NumberFormatException e) {}
         }
     }
